+++ src/common/utils/xmlUtils.ts
@@ -77,7 +77,7 @@ export async function isPackageXmlEmpty(
 export async function removePackageXmlFilesContent(
   packageXmlFile: string,
   removePackageXmlFile: string,
-  { outputXmlFile = null, logFlag = false, removedOnly = false }
+  { outputXmlFile = null, logFlag = false, removedOnly = false, keepEmptyTypes = false }
 ) {
   // Read package.xml file to update
   const parsedPackageXml: any = await parseXmlFile(packageXmlFile);
@@ -118,14 +118,30 @@ export async function removePackageXmlFilesContent(
     }
     const type = types[0];
     let typeMembers = type.members || [];
-    // Manage * case
-    if (removeTypeMembers[0] && removeTypeMembers[0] === "*") {
+    // Manage * case contained in target
+    if (removedOnly === true && typeMembers.includes("*")) {
+      typeMembers = removeTypeMembers;
+      uxLog(this, c.grey(c.italic(`Found wildcard * on type ${c.bold(type.name)}, kept values: ${typeMembers.join(",")}`)));
+    }
+    // Manage * case contained in source
+    else if (removeTypeMembers[0] && removeTypeMembers[0] === "*") {
       typeMembers = typeMembers.filter(() => checkRemove(false, removedOnly));
+      uxLog(this, c.grey(c.italic(`Found wildcard * on type ${c.bold(type.name)} which has been ${removedOnly ? "kept" : "removed"}`)));
     } else {
       // Filter members
       typeMembers = typeMembers.filter((member: string) => checkRemove(!removeTypeMembers.includes(member), removedOnly));
+      uxLog(
+        this,
+        c.grey(
+          c.italic(
+            `Found type ${c.bold(type.name)}, following elements has been ${removedOnly ? "removed" : "kept"}: ${
+              typeMembers.length > 0 ? typeMembers.join(",") : "none"
+            }`
+          )
+        )
+      );
     }
-    if (typeMembers.length > 0) {
+    if (typeMembers.length > 0 || keepEmptyTypes === true) {
       // Update members for type
       packageXmlMetadatasTypeLs = packageXmlMetadatasTypeLs.map((type1: any) => {
         if (type1.name[0] === type.name[0]) {